extend to flexible size of fields and added some tests

+++ arithmetic.c
@@ -2,22 +2,22 @@
 #include "arithmetic.h"
 
 // carry-less addition (use XOR)
-int gf_add_BCH_8bits(int x, int y)
+long gf_add_BCH_8bits(long x, long y)
 {
     return x ^ y;
 }
 
 // carry-less substraction (use XOR)
-int gf_sub_BCH_8bits(int x, int y)
+long gf_sub_BCH_8bits(long x, long y)
 {
     return x ^ y;
 }
 
 // multiplication
-int gf_mul_BCH_8bits(int x, int y)
+long gf_mul_BCH_8bits(long x, long y)
 {
-    int result = 0;
-    int pos = 0;
+    long result = 0;
+    long pos = 0;
     while ((y >> pos) > 0)
     {
         if (y & (1 << pos))
@@ -43,15 +43,15 @@ int gf_mul_BCH_8bits(int x, int y)
 // }
 
 // carry-less multiplication using the Russian Peasant Multiplication algorithm
-int gf_mul_MR_BCH(int x, int y, int prime_polynomial, int bits){
-    int result = 0;
+long gf_mul_MR_BCH(long x, long y, long prime_polynomial, long bits){
+    long result = 0;
     while (y > 0){
         if (y & 1){
             result = gf_add_BCH_8bits(result, x);
         }
         y = y >> 1;
         x = x << 1;
-        int temp = pow(2, bits);
+        long temp = pow(2, bits);
         if (prime_polynomial > 0 && (x & temp)){
             x = x ^ prime_polynomial;
         }
@@ -60,7 +60,7 @@ int gf_mul_MR_BCH(int x, int y, int prime_polynomial, int bits){
 }
 
 // multiplication using look up table
-int gf_mul_MR_BCH_8bits_LUT(int x, int y, struct Tables *tables)
+long gf_mul_MR_BCH_8bits_LUT(long x, long y, struct Tables *tables)
 {
     if (x == 0 || y == 0)
     {
@@ -70,7 +70,7 @@ int gf_mul_MR_BCH_8bits_LUT(int x, int y, struct Tables *tables)
 }
 
 // division using look up table
-int gf_div_MR_BCH_8bits_LUT(int x, int y, struct Tables *tables)
+long gf_div_MR_BCH_8bits_LUT(long x, long y, struct Tables *tables)
 {
     if (y == 0)
     {
@@ -81,39 +81,39 @@ int gf_div_MR_BCH_8bits_LUT(int x, int y, struct Tables *tables)
     {
         return 0;
     }
-    int log_size = tables->gf_log_size - 1;
+    long log_size = tables->gf_log_size - 1;
     return *(tables->gf_exp + (*(tables->gf_log + x) + log_size - *(tables->gf_log + y)) % log_size);
 }
 
 // power using look up table
-int gf_pow_MR_BCH_8bits_LUT(int x, int power, struct Tables *tables)
+long gf_pow_MR_BCH_8bits_LUT(long x, long power, struct Tables *tables)
 {
-    int log_size = tables->gf_log_size - 1;
+    long log_size = tables->gf_log_size - 1;
     if (power >= 0){
         return *(tables->gf_exp + (*(tables->gf_log + x) * power) % log_size);
     }else{
-        // printf("%d\n", (*(tables->gf_log + x) * power) % log_size + log_size);
+        // printf("%ld\n", (*(tables->gf_log + x) * power) % log_size + log_size);
         return *(tables->gf_exp + (*(tables->gf_log + x) * power) % log_size + log_size);
     }
 }
 
 // inverse using look up table
-int gf_inverse_MR_BCH_8bits_LUT(int x, struct Tables *tables)
+long gf_inverse_MR_BCH_8bits_LUT(long x, struct Tables *tables)
 {
-    int log_size = tables->gf_log_size - 1;
+    long log_size = tables->gf_log_size - 1;
     return *(tables->gf_exp + log_size - *(tables->gf_log + x));
 }
 
 // poly multiplication with integer: p = p * x, where p is a polynomial and x is an integer
-struct Polynomial *gf_poly_scale(struct Polynomial *p, int x, struct Tables *tables)
+struct Polynomial *gf_poly_scale(struct Polynomial *p, long x, struct Tables *tables)
 {
-    int arr[p->poly_size];
-    for (int i = 0; i < p->poly_size; i++)
+    long arr[p->poly_size];
+    for (long i = 0; i < p->poly_size; i++)
     {
         *(arr + i) = 0;
     }
     struct Polynomial *ret_val = newPolynomial(arr, p->poly_size);
-    for (int i = 0; i < p->poly_size; i++)
+    for (long i = 0; i < p->poly_size; i++)
     {
         *(ret_val->poly_arr + i) = gf_mul_MR_BCH_8bits_LUT(*(p->poly_arr + i), x, tables);
     }
@@ -123,7 +123,7 @@ struct Polynomial *gf_poly_scale(struct Polynomial *p, int x, struct Tables *tab
 // poly addition
 struct Polynomial *gf_poly_add(struct Polynomial *p, struct Polynomial *q)
 {
-    int ret_size = 0;
+    long ret_size = 0;
     if (p->poly_size > q->poly_size)
     {
         ret_size = p->poly_size;
@@ -137,7 +137,7 @@ struct Polynomial *gf_poly_add(struct Polynomial *p, struct Polynomial *q)
     }
     ret_val->poly_size = ret_size;
 
-    ret_val->poly_arr = malloc(ret_size * sizeof(int));
+    ret_val->poly_arr = malloc(ret_size * sizeof(long));
 
     if (ret_val->poly_arr == NULL)
     {
@@ -145,12 +145,12 @@ struct Polynomial *gf_poly_add(struct Polynomial *p, struct Polynomial *q)
         return NULL;
     }
 
-    for (int i = 0; i < p->poly_size; i++)
+    for (long i = 0; i < p->poly_size; i++)
     {
         *(ret_val->poly_arr + i + ret_size - p->poly_size) = *(p->poly_arr + i);
     }
 
-    for (int i = 0; i < q->poly_size; i++)
+    for (long i = 0; i < q->poly_size; i++)
     {
         *(ret_val->poly_arr + i + ret_size - q->poly_size) ^= *(q->poly_arr + i);
     }
@@ -168,9 +168,9 @@ struct Polynomial *gf_poly_mul(struct Polynomial *p, struct Polynomial *q, struc
     }
     ret_val->poly_size = (p->poly_size + q->poly_size - 1);
 
-    ret_val->poly_arr = malloc(ret_val->poly_size * sizeof(int));
+    ret_val->poly_arr = malloc(ret_val->poly_size * sizeof(long));
 
-    for (int i = 0; i < ret_val->poly_size; i++)
+    for (long i = 0; i < ret_val->poly_size; i++)
     {
         *(ret_val->poly_arr + i) = 0;
     }
@@ -181,9 +181,9 @@ struct Polynomial *gf_poly_mul(struct Polynomial *p, struct Polynomial *q, struc
         return NULL;
     }
 
-    for (int j = 0; j < q->poly_size; j++)
+    for (long j = 0; j < q->poly_size; j++)
     {
-        for (int i = 0; i < p->poly_size; i++)
+        for (long i = 0; i < p->poly_size; i++)
         {
             ret_val->poly_arr[i+j] = gf_add_BCH_8bits(ret_val->poly_arr[i+j],
             gf_mul_MR_BCH_8bits_LUT(*(p->poly_arr + i), *(q->poly_arr + j), tables));
@@ -194,10 +194,10 @@ struct Polynomial *gf_poly_mul(struct Polynomial *p, struct Polynomial *q, struc
 }
 
 // Evaluation: given x and the coefficients of a polynomial, convert it into a numeric number (integer)
-int gf_poly_eval(struct Polynomial *p, int x, struct Tables *tables)
+long gf_poly_eval(struct Polynomial *p, long x, struct Tables *tables)
 {
-    int y = *(p->poly_arr);
-    for (int i = 1; i < p->poly_size; i++)
+    long y = *(p->poly_arr);
+    for (long i = 1; i < p->poly_size; i++)
     {
         y = gf_mul_MR_BCH_8bits_LUT(y, x, tables) ^ *(p->poly_arr + i);
     }
@@ -205,11 +205,11 @@ int gf_poly_eval(struct Polynomial *p, int x, struct Tables *tables)
 }
 
 // rs generator: given number of symbols, generate an unreducible generator polynomial
-struct Polynomial *rs_generator_poly(int nsym, struct Tables *table){
-    int arr_g[1] = {1};
+struct Polynomial *rs_generator_poly(long nsym, struct Tables *table){
+    long arr_g[1] = {1};
     struct Polynomial *g = newPolynomial(arr_g, 1);
-    for (int i = 0; i < nsym; i++){
-        int arr_temp[2] = {1, gf_pow_MR_BCH_8bits_LUT(2, i, table)};
+    for (long i = 0; i < nsym; i++){
+        long arr_temp[2] = {1, gf_pow_MR_BCH_8bits_LUT(2, i, table)};
         struct Polynomial *temp = newPolynomial(arr_temp, 2);
         struct Polynomial *old_g = g;
         g = gf_poly_mul(g, temp, table);
@@ -221,13 +221,13 @@ struct Polynomial *rs_generator_poly(int nsym, struct Tables *table){
 
 // polynomial division, given a dividend and a divisor, fill the given pointer for qoutient and remainder.
 void gf_poly_div(struct Polynomial *qoutient, struct Polynomial *remainder, struct Polynomial *dividend, struct Polynomial *divisor, struct Tables *table){
-    int *msg_out = malloc(dividend->poly_size * sizeof(int));
-    memcpy(msg_out, dividend->poly_arr, dividend->poly_size * sizeof(int));
-    for (int i = 0; i < dividend->poly_size - divisor->poly_size + 1; i++){
-        int coef = msg_out[i];
+    long *msg_out = malloc(dividend->poly_size * sizeof(long));
+    memcpy(msg_out, dividend->poly_arr, dividend->poly_size * sizeof(long));
+    for (long i = 0; i < dividend->poly_size - divisor->poly_size + 1; i++){
+        long coef = msg_out[i];
         // check that coef is not 0
         if (coef){
-            for (int j = 1; j < divisor->poly_size; j++){
+            for (long j = 1; j < divisor->poly_size; j++){
                 if (*(divisor->poly_arr + j)){
                     msg_out[i + j] ^= gf_mul_MR_BCH_8bits_LUT(*(divisor->poly_arr + j), coef, table);
                 }
@@ -235,16 +235,16 @@ void gf_poly_div(struct Polynomial *qoutient, struct Polynomial *remainder, stru
         }
     }
     // separtor is the length of remainder (TODO: what if the length of remainder is 0?)
-    int separator = divisor->poly_size - 1;
+    long separator = divisor->poly_size - 1;
 
-    int *quo = malloc((dividend->poly_size - separator) * sizeof(int));
-    memcpy(quo, msg_out, (dividend->poly_size - separator) * sizeof(int));
+    long *quo = malloc((dividend->poly_size - separator) * sizeof(long));
+    memcpy(quo, msg_out, (dividend->poly_size - separator) * sizeof(long));
     // qoutient = newPolynomial(quo, dividend->poly_size - separator);
     qoutient->poly_size = dividend->poly_size - separator;
     qoutient->poly_arr = quo;
 
-    int *remain = malloc(sizeof(int) * separator);
-    memcpy(remain, msg_out + dividend->poly_size - separator, separator * sizeof(int));
+    long *remain = malloc(sizeof(long) * separator);
+    memcpy(remain, msg_out + dividend->poly_size - separator, separator * sizeof(long));
     // remainder = newPolynomial(remain, separator);
     remainder->poly_size = separator;
     remainder->poly_arr = remain;
