+++ utility.h
@@ -3,25 +3,25 @@
 #include <math.h>
 #include <string.h>
 
-void printBinary(int);
+void printBinary(long);
 
 void reverse_string(char *);
 
-int bit_length(int);
+long bit_length(long);
 
-int carry_less_long_div(int, int);
+long carry_less_long_div(long, long);
 
-int gf_mul_MR_BCH(int, int, int, int);
+long gf_mul_MR_BCH(long, long, long, long);
 
 struct Tables
 {
-    int *gf_exp;
+    long *gf_exp;
     size_t gf_exp_size;
-    int *gf_log;
+    long *gf_log;
     size_t gf_log_size;
 };
 
-struct Tables *newTables(int, size_t);
+struct Tables *newTables(long, long);
 
 void delTables(struct Tables *);
 
@@ -30,10 +30,10 @@ void printTables(struct Tables *);
 struct Polynomial
 {
     size_t poly_size;
-    int *poly_arr;
+    long *poly_arr;
 };
 
-struct Polynomial *newPolynomial(int *, size_t);
+struct Polynomial *newPolynomial(long *, size_t);
 
 void delPolynomial(struct Polynomial *);
 
@@ -41,19 +41,19 @@ struct Polynomial* reversePolynomial(struct Polynomial *);
 
 void printPolynomial(struct Polynomial *poly);
 
-void printPolynomialAsMessage(struct Polynomial *poly, int k);
+void printPolynomialAsMessage(struct Polynomial *poly, long k);
 
 struct DynamicArray
 {
     size_t arr_size;
     size_t capacity;
-    int *data;
+    long *data;
 };
 
 struct DynamicArray *newDynamicArray(size_t);
 
 void delDynamicArray(struct DynamicArray *);
 
-void push_back(struct DynamicArray *, int);
+void push_back(struct DynamicArray *, long);
 
 void printDynamicArray(struct DynamicArray *);
